Fix bug where scaler only went up

+++ lib/serving_scaler.dart
@@ -5,7 +5,7 @@ class ServingScaler {
 
   ServingScaler(this.ingredients);
 
-  List<Ingredient> amountScaler(int servings) {
+  List<Ingredient> amountScaler(double servings) {
     for (int i = 0; i < ingredients.length; i++) {
       double amount = ingredients[i].amount;
       amount = amount * servings;
